extracted logic into override config package
